Improve paradrop to keep aircraft level
Placing a waypoint for the drop zone was a problem because the
aircraft would turn to the next waypoint, causing it to be banked to
the side when units drop. Now, instead of a waypoint, the drop code
happens when the aircraft is 40% between the ingress and egress
waypoints, and it will be level and flying straight when the drop
happens.